cleanup

+++ matlab/convert_spikesmatrix.py
@@ -24,7 +24,7 @@ def convert(num_pn,inp='./spikesmatrix_op_ryt', out='../vpsi_inh_spikes.h5'):
         if len(times)==1 and times[0]=='':
             count = count + 1
             continue # There were no spikes for this cell
-        times = [float(time) for time in times if float(time) < maxtime]
+        times = [int(time) for time in times if int(time) < maxtime]
         times.sort()
         ids = [count for _ in range(len(times))]
